chore: extract setupSteps state into context provider
